+++ lib/xe_client/errors/authentication_error.rb
@@ -1,0 +1,4 @@
+module XEClient
+  class AuthenticationError < Error
+  end
+end