+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- Raise XEClient::AuthenticationError when there's an auth issue
+- Raise XEClient::Error for all other errors
+
 ## [0.1.0] - 2016-06-22
 ### Added
 - Add initial working version (`#convert_from`)